+++ app/ante/evm/ante_test.go
@@ -42,7 +42,7 @@ func (suite *AnteTestSuite) TestAnteHandler() {
 		ctx = suite.GetNetwork().GetContext()
 	}
 
-	evmChainID := evmtypes.GetChainConfig().ChainID
+	evmChainID := evmtypes.GetEthChainConfig().ChainID
 	ethContractCreationTxParams := evmtypes.EvmTxArgs{
 		ChainID:   evmChainID,
 		Nonce:     0,
@@ -166,7 +166,7 @@ func (suite *AnteTestSuite) TestAnteHandler() {
 				nonce, err := suite.GetNetwork().App.AccountKeeper.GetSequence(ctx, suite.GetKeyring().GetAccAddr(0))
 				suite.Require().NoError(err)
 				ethTxParams := evmtypes.EvmTxArgs{
-					ChainID:  evmtypes.GetChainConfig().ChainID,
+					ChainID:  evmtypes.GetEthChainConfig().ChainID,
 					To:       &to,
 					Nonce:    nonce,
 					Amount:   big.NewInt(10),
@@ -185,7 +185,7 @@ func (suite *AnteTestSuite) TestAnteHandler() {
 				nonce, err := suite.GetNetwork().App.AccountKeeper.GetSequence(ctx, suite.GetKeyring().GetAccAddr(0))
 				suite.Require().NoError(err)
 				ethTxParams := evmtypes.EvmTxArgs{
-					ChainID:  evmtypes.GetChainConfig().ChainID,
+					ChainID:  evmtypes.GetEthChainConfig().ChainID,
 					To:       &to,
 					Nonce:    nonce,
 					Amount:   big.NewInt(10),
@@ -203,7 +203,7 @@ func (suite *AnteTestSuite) TestAnteHandler() {
 				nonce, err := suite.GetNetwork().App.AccountKeeper.GetSequence(ctx, suite.GetKeyring().GetAccAddr(0))
 				suite.Require().NoError(err)
 				ethTxParams := evmtypes.EvmTxArgs{
-					ChainID:  evmtypes.GetChainConfig().ChainID,
+					ChainID:  evmtypes.GetEthChainConfig().ChainID,
 					To:       &to,
 					Nonce:    nonce,
 					Amount:   big.NewInt(10),
@@ -221,7 +221,7 @@ func (suite *AnteTestSuite) TestAnteHandler() {
 				nonce, err := suite.GetNetwork().App.AccountKeeper.GetSequence(ctx, suite.GetKeyring().GetAccAddr(0))
 				suite.Require().NoError(err)
 				ethTxParams := evmtypes.EvmTxArgs{
-					ChainID:  evmtypes.GetChainConfig().ChainID,
+					ChainID:  evmtypes.GetEthChainConfig().ChainID,
 					To:       &to,
 					Nonce:    nonce,
 					Amount:   big.NewInt(10),
@@ -243,7 +243,7 @@ func (suite *AnteTestSuite) TestAnteHandler() {
 				nonce, err := suite.GetNetwork().App.AccountKeeper.GetSequence(ctx, suite.GetKeyring().GetAccAddr(0))
 				suite.Require().NoError(err)
 				ethTxParams := evmtypes.EvmTxArgs{
-					ChainID:  evmtypes.GetChainConfig().ChainID,
+					ChainID:  evmtypes.GetEthChainConfig().ChainID,
 					To:       &to,
 					Nonce:    nonce,
 					Amount:   big.NewInt(10),
@@ -920,7 +920,7 @@ func (suite *AnteTestSuite) TestAnteHandler() {
 func (suite *AnteTestSuite) TestAnteHandlerWithDynamicTxFee() {
 	addr, privKey := utiltx.NewAddrKey()
 	to := utiltx.GenerateAddress()
-	evmChainID := evmtypes.GetChainConfig().ChainID
+	evmChainID := evmtypes.GetEthChainConfig().ChainID
 
 	ethContractCreationTxParams := evmtypes.EvmTxArgs{
 		ChainID:   evmChainID,
@@ -1084,7 +1084,7 @@ func (suite *AnteTestSuite) TestAnteHandlerWithDynamicTxFee() {
 func (suite *AnteTestSuite) TestAnteHandlerWithParams() {
 	addr, privKey := utiltx.NewAddrKey()
 	to := utiltx.GenerateAddress()
-	evmChainID := evmtypes.GetChainConfig().ChainID
+	evmChainID := evmtypes.GetEthChainConfig().ChainID
 
 	ethContractCreationTxParams := evmtypes.EvmTxArgs{
 		ChainID:   evmChainID,