+++ app/ante/testutils/testutils.go
@@ -106,7 +106,7 @@ func (suite *AnteTestSuite) SetupTest() {
 	denom := evmtypes.GetEVMCoinDenom()       //nolint:staticcheck
 	decimals := evmtypes.GetEVMCoinDecimals() //nolint:staticcheck
 	configurator := evmtypes.NewEVMConfigurator()
-	configurator.ResetTestChainConfig()
+	configurator.ResetTestConfig()
 	err := configurator.
 		WithChainConfig(chainConfig).
 		WithEVMCoinInfo(denom, uint8(decimals)).