+++ x/evm/types/chain_config.go
@@ -23,7 +23,7 @@ const testChainID string = "evmos_9002-1"
 
 // chainConfig is the chain configuration used in the EVM to defined which
 // opcodes are active based on Ethereum upgrades.
-var chainConfig *geth.ChainConfig
+var chainConfig *ChainConfig
 
 // EthereumConfig returns an Ethereum ChainConfig for EVM state transitions.
 // All the negative or nil values are converted to nil
@@ -125,7 +125,7 @@ func setChainConfig(cc *ChainConfig) error {
 	if err := config.Validate(); err != nil {
 		return err
 	}
-	chainConfig = config.EthereumConfig(nil)
+	chainConfig = config
 
 	return nil
 }