+++ x/evm/types/config.go
@@ -4,22 +4,17 @@
 // The config package provides a convenient way to modify x/evm params and values.
 // Its primary purpose is to be used during application initialization.
 
-//go:build test
-// +build test
+//go:build !test
+// +build !test
 
 package types
 
 import (
-	"errors"
 	"fmt"
 
 	geth "github.com/ethereum/go-ethereum/params"
 	"github.com/evmos/evmos/v20/x/evm/core/vm"
 )
-
-// testChainConfig is the chain configuration used in the EVM to defined which
-// opcodes are active based on Ethereum upgrades.
-var testChainConfig *ChainConfig
 
 // Configure applies the changes to the virtual machine configuration.
 func (ec *EVMConfigurator) Configure() error {
@@ -29,11 +24,11 @@ func (ec *EVMConfigurator) Configure() error {
 		return fmt.Errorf("error configuring EVMConfigurator: already sealed and cannot be modified")
 	}
 
-	if err := setTestChainConfig(ec.chainConfig); err != nil {
+	if err := setChainConfig(ec.chainConfig); err != nil {
 		return err
 	}
 
-	if err := setTestingEVMCoinInfo(ec.evmCoinInfo); err != nil {
+	if err := setEVMCoinInfo(ec.evmCoinInfo); err != nil {
 		return err
 	}
 
@@ -45,7 +40,7 @@ func (ec *EVMConfigurator) Configure() error {
 		return err
 	}
 
-	// After applying modifications, the configurator is sealed. This way, it is not possible
+	// After applying modifiers the configurator is sealed. This way, it is not possible
 	// to call the configure method twice.
 	ec.sealed = true
 
@@ -53,32 +48,15 @@ func (ec *EVMConfigurator) Configure() error {
 }
 
 func (ec *EVMConfigurator) ResetTestConfig() {
-	vm.ResetActivators()
-	resetEVMCoinInfo()
-	testChainConfig = nil
-}
-
-func setTestChainConfig(cc *ChainConfig) error {
-	if testChainConfig != nil {
-		return errors.New("chainConfig already set. Cannot set again the chainConfig. Call the configurators ResetTestConfig method before configuring a new chain.")
-	}
-	config := DefaultChainConfig("")
-	if cc != nil {
-		config = cc
-	}
-	if err := config.Validate(); err != nil {
-		return err
-	}
-	testChainConfig = config
-	return nil
+	panic("this is only implemented with the 'test' build flag. Make sure you're running your tests using the '-tags=test' flag.")
 }
 
 // GetEthChainConfig returns the `chainConfig` used in the EVM (geth type).
 func GetEthChainConfig() *geth.ChainConfig {
-	return testChainConfig.EthereumConfig(nil)
+	return chainConfig.EthereumConfig(nil)
 }
 
 // GetChainConfig returns the `chainConfig`.
 func GetChainConfig() *ChainConfig {
-	return testChainConfig
+	return chainConfig
 }